removed codesmells
